Fixed Array Length Problems
Previously had the user specified the array size '-a' to be less than 1 the program returns that it sorted too fast. Now filters for numbers less than or equal to 1 and tells the user to specify a larger number. 
Also, improved the system for checking if the array is in order, previously checked whole numbers, and now checks if the current number in the array is greater than the last. Mostly in preparation for future changes.