Added support for parallel processing where possible.
While not all sorting algorithms will be at an advantage when using
parallel processing, many can benefit if they do so correctly. Now, when
the user specifies '-p' in the program parameters it will use all
available CPU resources to sort!

Yes I know this could cause problems for some algorithms, which is why
you have the option not to use it. But concurrency should assure that no
real problems occur as variables can be externally shared.