Made changes to parallel system in SortingAlgorithm
Made some small changes to the system. It should be noted that with the
current system and Operating System controlled threads there is a
possibility that for some OS's. I know Windows for sure and depending on
your processing hardware parallel processing may have little or no
effect or only occasionally provide any marginal performance
improvements. It is most useful and consistent in showing improvement on
servers, expeciallt those running a Linux based system such as Ubuntu
Server.